+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/config/MetadataTransferAutoConfiguration.java
@@ -26,6 +26,7 @@ import com.tencent.cloud.metadata.core.EncodeTransferMedataRestTemplateEnhancedP
 import com.tencent.cloud.metadata.core.EncodeTransferMedataScgEnhancedPlugin;
 import com.tencent.cloud.metadata.core.EncodeTransferMedataWebClientEnhancedPlugin;
 import com.tencent.cloud.metadata.core.EncodeTransferMetadataZuulEnhancedPlugin;
+import com.tencent.cloud.polaris.context.config.PolarisContextProperties;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -66,8 +67,8 @@ public class MetadataTransferAutoConfiguration {
 		}
 
 		@Bean
-		public DecodeTransferMetadataServletFilter metadataServletFilter() {
-			return new DecodeTransferMetadataServletFilter();
+		public DecodeTransferMetadataServletFilter metadataServletFilter(PolarisContextProperties polarisContextProperties) {
+			return new DecodeTransferMetadataServletFilter(polarisContextProperties);
 		}
 	}
 
@@ -79,8 +80,8 @@ public class MetadataTransferAutoConfiguration {
 	protected static class MetadataReactiveFilterConfig {
 
 		@Bean
-		public DecodeTransferMetadataReactiveFilter metadataReactiveFilter() {
-			return new DecodeTransferMetadataReactiveFilter();
+		public DecodeTransferMetadataReactiveFilter metadataReactiveFilter(PolarisContextProperties polarisContextProperties) {
+			return new DecodeTransferMetadataReactiveFilter(polarisContextProperties);
 		}
 
 	}