+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/DecodeTransferMetadataServletFilter.java
@@ -32,6 +32,7 @@ import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.util.JacksonUtils;
 import com.tencent.cloud.common.util.UrlUtils;
 import com.tencent.cloud.metadata.provider.ServletMetadataProvider;
+import com.tencent.cloud.polaris.context.config.PolarisContextProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,6 +54,12 @@ public class DecodeTransferMetadataServletFilter extends OncePerRequestFilter {
 
 	private static final Logger LOG = LoggerFactory.getLogger(DecodeTransferMetadataServletFilter.class);
 
+	private PolarisContextProperties polarisContextProperties;
+
+	public DecodeTransferMetadataServletFilter(PolarisContextProperties polarisContextProperties) {
+		this.polarisContextProperties = polarisContextProperties;
+	}
+
 	@Override
 	protected void doFilterInternal(@NonNull HttpServletRequest httpServletRequest,
 			@NonNull HttpServletResponse httpServletResponse, FilterChain filterChain)
@@ -65,7 +72,7 @@ public class DecodeTransferMetadataServletFilter extends OncePerRequestFilter {
 		mergedTransitiveMetadata.putAll(customTransitiveMetadata);
 		Map<String, String> internalDisposableMetadata = getInternalMetadata(httpServletRequest, CUSTOM_DISPOSABLE_METADATA);
 		Map<String, String> mergedDisposableMetadata = new HashMap<>(internalDisposableMetadata);
-		ServletMetadataProvider metadataProvider = new ServletMetadataProvider(httpServletRequest);
+		ServletMetadataProvider metadataProvider = new ServletMetadataProvider(httpServletRequest, polarisContextProperties.getLocalIpAddress());
 		MetadataContextHolder.init(mergedTransitiveMetadata, mergedDisposableMetadata, metadataProvider);
 
 		TransHeadersTransfer.transfer(httpServletRequest);