+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/config/MetadataTransferAutoConfigurationTest.java
@@ -21,6 +21,7 @@ package com.tencent.cloud.metadata.config;
 import com.tencent.cloud.metadata.core.EncodeTransferMedataFeignEnhancedPlugin;
 import com.tencent.cloud.metadata.core.EncodeTransferMedataRestTemplateEnhancedPlugin;
 import com.tencent.cloud.metadata.core.EncodeTransferMedataWebClientEnhancedPlugin;
+import com.tencent.cloud.polaris.context.config.PolarisContextProperties;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -63,7 +64,7 @@ public class MetadataTransferAutoConfigurationTest {
 	 */
 	@Test
 	public void test2() {
-		this.reactiveWebApplicationContextRunner.withConfiguration(AutoConfigurations.of(MetadataTransferAutoConfiguration.class))
+		this.reactiveWebApplicationContextRunner.withConfiguration(AutoConfigurations.of(MetadataTransferAutoConfiguration.class, PolarisContextProperties.class))
 				.run(context -> {
 					assertThat(context).hasSingleBean(MetadataTransferAutoConfiguration.MetadataTransferFeignInterceptorConfig.class);
 					assertThat(context).hasSingleBean(EncodeTransferMedataFeignEnhancedPlugin.class);