+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [feat: PolarisServiceRegistry#deregister support idempotency.](https://github.com/Tencent/spring-cloud-tencent/pull/1243)
 - [feat: SCT元数据管理能力与polaris-java元数据管理能力进行下沉及整合](https://github.com/Tencent/spring-cloud-tencent/pull/1249)
 - [feat: support lane router](https://github.com/Tencent/spring-cloud-tencent/pull/1250)
+- [feat: lane router rule support caller ip](https://github.com/Tencent/spring-cloud-tencent/pull/1253)